:recycle: Seperate Category footer
