+++ src/app/dashboard/_components/Category.tsx
@@ -2,7 +2,6 @@ import {
   CheckOutlined,
   DeleteOutlined,
   EditOutlined,
-  PlusOutlined,
   MoreOutlined,
 } from "@ant-design/icons";
 import { Input, Select } from "antd";
@@ -23,6 +22,7 @@ import useClickOutside from "../_hooks/useClickOutside";
 import Card from "./Card";
 import useDragItem from "../_hooks/dnd/useDragItem";
 import useDropItem from "../_hooks/dnd/useDropItem";
+import CategoryFooter from "./CategoryFooter";
 
 function Category({
   index,
@@ -44,21 +44,17 @@ function Category({
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
   const [selectedPosition, setSelectedPosition] = useState(index + 1);
   const [cardList, setCardList] = useState<ICard[]>([]);
-  const [isAddingCard, setIsAddingCard] = useState(false);
-  const [newCardTitle, setNewCardTitle] = useState("");
 
   const dragRef = useRef<HTMLDivElement>(null);
   const editTitleInputRef = useRef<HTMLDivElement>(null);
   const moreModalRef = useRef<HTMLDivElement>(null);
   const copyModalRef = useRef<HTMLDivElement>(null);
   const moveModalRef = useRef<HTMLDivElement>(null);
-  const addCardRef = useRef<HTMLDivElement>(null);
 
   useClickOutside(setIsEditingTitle, editTitleInputRef);
   useClickOutside(setIsMoreVisible, moreModalRef);
   useClickOutside(setIsCopyModalVisible, copyModalRef);
   useClickOutside(setIsMoveModalVisible, moveModalRef, isDropdownOpen);
-  useClickOutside(setIsAddingCard, addCardRef);
 
   const { isDragging, drag } = useDragItem("category", category.id, index);
   const { handlerId, drop } = useDropItem(dragRef, index, onDragCategory);
@@ -142,22 +138,11 @@ function Category({
     setIsMoveModalVisible(false);
   };
 
-  const handleAddCard = async () => {
-    setIsAddingCard(true);
-  };
-
-  const handleSaveCard = async () => {
+  const onSaveCard = async (newCardTitle: string) => {
     if (newCardTitle.trim() === "" || !cardRepository) return;
 
     await cardRepository.add(newCardTitle, category.id);
     fetchCards();
-    setIsAddingCard(false);
-    setNewCardTitle("");
-  };
-
-  const handleCancelSaveCard = () => {
-    setIsAddingCard(false);
-    setNewCardTitle("");
   };
 
   return (
@@ -287,42 +272,7 @@ function Category({
           ))}
         </div>
       </div>
-      <footer className="p-2">
-        <button
-          type="button"
-          className="flex items-center gap-2 text-sm bg-sky-300 text-white px-2 py-1 rounded-lg hover:bg-sky-400"
-          onClick={handleAddCard}
-        >
-          <PlusOutlined />
-          <div> Add Card</div>
-        </button>
-        {isAddingCard && (
-          <div className="flex flex-col gap-2 mt-2" ref={addCardRef}>
-            <Input
-              placeholder="Card name"
-              autoFocus
-              onChange={(e) => setNewCardTitle(e.target.value)}
-              value={newCardTitle}
-            />
-            <div className="flex justify-center items-center gap-2">
-              <button
-                type="button"
-                className="bg-emerald-400 text-white rounded-lg px-2 py-1 font-semibold"
-                onClick={handleSaveCard}
-              >
-                Save
-              </button>
-              <button
-                type="button"
-                className="bg-rose-600 text-white rounded-lg px-2 py-1 font-semibold"
-                onClick={handleCancelSaveCard}
-              >
-                Cancel
-              </button>
-            </div>
-          </div>
-        )}
-      </footer>
+      <CategoryFooter onSaveCard={onSaveCard} />
     </div>
   );
 }
